+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Features
 
+- Set value to immutable field from another field.
+
 ## Bug fixes
 
 ## TODO